+++ Modules/Accounting/Entities/SignedMoneyTransaction.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Modules\Accounting\Entities;
+
+use Illuminate\Database\Eloquent\Model;
+
+class SignedMoneyTransaction extends Model
+{
+    protected $table = 'accounting_signed_transactions';
+}